поправил некоторые ошибки по тэгам v2

+++ src/main/java/ru/skillbox/zerone/backend/service/TagService.java
@@ -5,6 +5,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import ru.skillbox.zerone.backend.exception.TagNotFoundException;
 import ru.skillbox.zerone.backend.mapstruct.TagMapper;
 import ru.skillbox.zerone.backend.model.dto.request.TagDTO;
 import ru.skillbox.zerone.backend.model.dto.response.CommonListResponseDTO;
@@ -13,11 +14,7 @@ import ru.skillbox.zerone.backend.model.dto.response.MessageResponseDTO;
 import ru.skillbox.zerone.backend.model.entity.Tag;
 import ru.skillbox.zerone.backend.repository.TagRepository;
 import ru.skillbox.zerone.backend.util.ResponseUtils;
-
-import java.security.NoSuchAlgorithmException;
-import java.security.SecureRandom;
 import java.util.List;
-import java.util.Random;
 
 @Service
 @RequiredArgsConstructor
@@ -29,10 +26,14 @@ public class TagService {
   @Transactional
   public CommonResponseDTO<TagDTO> addTag(TagDTO tagDTO) {
 
-    Tag tag = tagMapper.tagDTOToTag(tagDTO);
+    if (tagDTO.getId()==null) {
+      String tagName = tagDTO.getTag();
+      Tag tag = tagRepository.findByName(tagName);
+      tagDTO.setId(tag.getId());
+    }
 
+    Tag tag = tagMapper.tagDTOToTag(tagDTO);
     tagRepository.save(tag);
-    tagDTO.setId(-1L);
 
     return CommonResponseDTO.<TagDTO>builder()
         .data(tagDTO)
@@ -41,6 +42,10 @@ public class TagService {
 
   @Transactional
   public CommonResponseDTO<MessageResponseDTO> deleteTag (Long id) {
+
+    if (id==null) {
+        throw new TagNotFoundException(id);
+    }
 
     tagRepository.deleteById(id);
 
@@ -55,7 +60,7 @@ public class TagService {
     if (tagName.isEmpty()) {
       pageableTagList = tagRepository.findAll(pageable);
     } else {
-      pageableTagList = tagRepository.findByName(tagName, pageable);
+      pageableTagList = tagRepository.findByNameContains(tagName, pageable);
     }
 
     List<TagDTO> tagDTOList = pageableTagList.map(tagMapper::tagToTagDTO).toList();
